+++ lib/node/index.d.ts
@@ -1,3 +1,3 @@
-export { PseudoStateKind, State, Region, PseudoState, ExternalTransition, InternalTransition, LocalTransition } from './model';
+export { PseudoStateKind, State, Region, PseudoState, Transition } from './model';
 export { IInstance, IRegion, IState, Instance } from './runtime';
 export { log, random } from './util';