+++ lib/node/model/index.d.ts
@@ -5,6 +5,3 @@ export { State } from './State';
 export { Region } from './Region';
 export { PseudoState } from './PseudoState';
 export { Transition } from './Transition';
-export { ExternalTransition } from './ExternalTransition';
-export { InternalTransition } from './InternalTransition';
-export { LocalTransition } from './LocalTransition';