+++ src/index.ts
@@ -1,4 +1,4 @@
-export { PseudoStateKind, State, Region, PseudoState, ExternalTransition, InternalTransition, LocalTransition } from './model';
+export { PseudoStateKind, State, Region, PseudoState, Transition } from './model';
 
 export { IInstance, IRegion, IState, Instance } from './runtime';