+++ src/model/index.ts
@@ -7,6 +7,6 @@ export { Region } from './Region';
 export { PseudoState } from './PseudoState';
 
 export { Transition } from './Transition';
-export { ExternalTransition } from './ExternalTransition';
-export { InternalTransition } from './InternalTransition';
-export { LocalTransition } from './LocalTransition';
+//export { ExternalTransition } from './ExternalTransition';
+//export { InternalTransition } from './InternalTransition';
+//export { LocalTransition } from './LocalTransition';